microservices/bAnalyzePicture/requirements.txt
google-cloud-pubsub==1.0.2
google-cloud-storage==1.23.0
google-cloud-vision==0.40.0
flask==1.1.1
python-dotenv==0.10.3